requirements.txt
langgraph
langchain_anthropic
tavily-python
langchain_community
langchain_openai
langchain_groq
pyowm
python-dotenv
setuptools # Required by pyowm for pkg_resources
langchain-huggingface
langchain-chroma